+++ src/main/java/io/github/matheusbeoulve/icecreamdemo/domain/Inventory.java
@@ -6,6 +6,7 @@ import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Table(name = "inventory")
@@ -30,7 +31,7 @@ public class Inventory {
     private Flavor flavor;
 
     @Column(nullable = false)
-    private Long stock;
+    private BigDecimal stock;
 
     @CreationTimestamp
     private LocalDateTime creation;
